Document what happened to most Rules.
This an initial attempt at this. I think it contains enough detail that a developer audience should be able to search through the code or UI (in the case of payment gateways) and figure out what to do from there. I could _perhaps_ say a bit more about how to add new payment gateway conditions, but I think that should be documented elsewhere and perhaps linked to from here, like resolvers.