Note that receipt field accepts multiple addresses (#163)
Not sure if that is the intention as a multi-value field would probably be better UX if multiples are intended.

Apparently 'types of orders' is better than 'types of order'. Maybe the singular is an NZ thing, or maybe it's just me? In any case the plural reads well to me.